+++ lib/spill.rb
@@ -8,6 +8,7 @@ require_relative "spill/collectors/github"
 require_relative "spill/report"
 require_relative "spill/renderer"
 require_relative "spill/spinner"
+require_relative "spill/narrator"
 require_relative "spill/cli"
 
 module Spill